refactor(rootfs): move virtiofs mount logic from vm to rootfs module
- Move update_rootfs_fstab from MicroVm to rootfs module and rename to patch_rootfs_with_virtiofs_mounts
- Move VIRTIOFS_TAG_PREFIX constant to vm module for reuse
- Rename clear_and_add_scripts_to_dir to patch_rootfs_with_sandbox_scripts for clarity
- Remove unused imports and dependencies from rootfs.rs
- Move associated tests from vm.rs to rootfs.rs
- Update sandbox.rs to use renamed function
