+++ virtualfs/examples/overlaynfs.rs
@@ -13,13 +13,16 @@
 //!
 //! To run the example:
 //! ```bash
-//! cargo run --example nfs -- /path/to/store --port 2049
+//! cargo run --example overlaynfs /path/to/layer1 /path/to/layer2 --host 127.0.0.1 --port 2049
 //! ```
 
 use anyhow::Result;
 use clap::Parser;
-use std::path::PathBuf;
-use virtualfs::{MemoryFileSystem, VirtualFileSystemServer, DEFAULT_HOST, DEFAULT_NFS_PORT};
+use std::{net::IpAddr, path::PathBuf};
+use virtualfs::{
+    NativeFileSystem, OverlayFileSystem, VirtualFileSystem, VirtualFileSystemServer,
+    DEFAULT_NFS_HOST, DEFAULT_NFS_PORT,
+};
 
 //--------------------------------------------------------------------------------------------------
 // Types
@@ -33,8 +36,8 @@ struct Args {
     layers: Vec<PathBuf>,
 
     /// Host address to bind to
-    #[arg(short = 'H', long, default_value = DEFAULT_HOST)]
-    host: String,
+    #[arg(short = 'H', long, default_value = DEFAULT_NFS_HOST)]
+    host: IpAddr,
 
     /// Port to listen on
     #[arg(short = 'P', long, default_value_t = DEFAULT_NFS_PORT)]
@@ -53,12 +56,19 @@ async fn main() -> Result<()> {
     // Initialize logging
     tracing_subscriber::fmt::init();
 
+    // Create filesystem layers from the provided paths
+    let mut layers: Vec<Box<dyn VirtualFileSystem + Send + Sync>> = Vec::new();
+
+    for path in &args.layers {
+        let fs = Box::new(NativeFileSystem::new(path.clone()));
+        layers.push(fs);
+    }
+
     // Create and start the server
-    // let fs = OverlayFileSystem::new()?;
-    let fs = MemoryFileSystem::new();
-    let server = VirtualFileSystemServer::new(fs, args.host, args.port);
+    let fs = OverlayFileSystem::new(layers)?;
+    let server = VirtualFileSystemServer::new(fs, args.host.to_string(), args.port);
     tracing::info!(
-        "Starting NFS server on {}:{}",
+        "starting NFS server on {}:{}",
         server.get_host(),
         server.get_port()
     );
