+++ virtualfs/lib/defaults.rs
@@ -3,7 +3,7 @@
 //--------------------------------------------------------------------------------------------------
 
 /// The default host address to bind to.
-pub const DEFAULT_HOST: &str = "127.0.0.1";
+pub const DEFAULT_NFS_HOST: &str = "127.0.0.1";
 
 /// The default NFS port number to use.
 pub const DEFAULT_NFS_PORT: u32 = 2049;
